requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
faiss-cpu==1.6.5
fastapi==0.63.0
filelock==3.0.12
h11==0.12.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.5
packaging==20.8
pandas==1.2.0
Pillow==8.1.0
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
Sastrawi==1.0.1
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
starlette==0.13.6
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
tqdm==4.56.0
transformers==4.1.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
wordcloud==1.8.1
